Remove NoopJournal
